Use multisite refactor of Ema, and new unionmount
Split sites into two.
