+++ ema-template.cabal
@@ -32,7 +32,7 @@ executable ema-template
     , data-default
     , dependent-sum
     , directory
-    , ema                 >=0.6
+    , ema                 >=0.7
     , filepath
     , lvar
     , monad-logger
@@ -103,5 +103,9 @@ executable ema-template
     ViewPatterns
 
   main-is:            Main.hs
+  other-modules:
+    Site.MarkdownSite
+    Site.StaticSite
+
   hs-source-dirs:     src
   default-language:   Haskell2010
